Web_Development/requirements.txt
absl-py==2.1.0
altair==5.5.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
gitpython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.29.2
idna==3.10
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
libclang==18.1.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.30.0
networkx==3.4.2
numpy==2.0.2
opt-einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0.1
protobuf==5.29.3
pyarrow==19.0.1
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
streamlit==1.43.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
timm==1.0.15
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.5
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
